# Allow lists to be built from iterators