--- a/rustler/src/types/list.rs
+++ b/rustler/src/types/list.rs
@@ -2,7 +2,7 @@
 //!
 //! Right now the only supported way to read lists are through the ListIterator.
 
-use crate::wrapper::{list, NIF_TERM};
+use crate::wrapper::list;
 use crate::{Decoder, Encoder, Env, Error, NifResult, Term};
 
 /// Enables iteration over the items in the list.
@@ -86,13 +86,6 @@ impl<'a> Decoder<'a> for ListIterator<'a> {
     }
 }
 
-//impl<'a, T> Encoder for Iterator<Item = T> where T: Encoder {
-//    fn encode<'b>(&self, env: Env<'b>) -> Term<'b> {
-//        let term_arr: Vec<NIF_TERM> =
-//            self.map(|x| x.encode(env).as_c_arg()).collect();
-//    }
-//}
-
 impl<T> Encoder for Vec<T>
 where
     T: Encoder,
@@ -118,8 +111,8 @@ where
     T: Encoder,
 {
     fn encode<'b>(&self, env: Env<'b>) -> Term<'b> {
-        let term_array: Vec<NIF_TERM> = self.iter().map(|x| x.encode(env).as_c_arg()).collect();
-        unsafe { Term::new(env, list::make_list(env.as_c_arg(), &term_array)) }
+        let mut iter = self.iter().map(|x| x.encode(env).as_c_arg());
+        unsafe { Term::new(env, list::make_list_from_end(env.as_c_arg(), &mut iter)) }
     }
 }
 impl<'a, T> Encoder for &'a [T]
@@ -127,8 +120,8 @@ where
     T: Encoder,
 {
     fn encode<'b>(&self, env: Env<'b>) -> Term<'b> {
-        let term_array: Vec<NIF_TERM> = self.iter().map(|x| x.encode(env).as_c_arg()).collect();
-        unsafe { Term::new(env, list::make_list(env.as_c_arg(), &term_array)) }
+        let mut iter = self.iter().map(|x| x.encode(env).as_c_arg());
+        unsafe { Term::new(env, list::make_list_from_end(env.as_c_arg(), &mut iter)) }
     }
 }