--- a/rustler_tests/native/rustler_test/src/test_list.rs
+++ b/rustler_tests/native/rustler_test/src/test_list.rs
@@ -1,5 +1,5 @@
-use rustler::{Encoder, Env, Error, ListIterator, NifResult, Term};
 use rustler::wrapper::list;
+use rustler::{Encoder, Env, Error, ListIterator, NifResult, Term};
 
 #[rustler::nif]
 pub fn sum_list(iter: ListIterator) -> NifResult<i64> {
@@ -16,15 +16,14 @@ pub fn make_list() -> Vec<usize> {
     vec![1, 2, 3]
 }
 
-
 #[rustler::nif]
 pub fn make_list_from_iter(env: Env) -> Term {
-  let mut iter = [1, 2, 3].iter().map(|i| (i + 1).encode(env).as_c_arg());
-  unsafe { Term::new(env, list::make_list_from_iter(env.as_c_arg(), &mut iter)) }
+    let mut iter = [1, 2, 3].iter().map(|i| (i + 1).encode(env).as_c_arg());
+    unsafe { Term::new(env, list::make_list_from_iter(env.as_c_arg(), &mut iter)) }
 }
 
 #[rustler::nif]
 pub fn make_list_from_end(env: Env) -> Term {
-  let mut iter = [1, 2, 3].iter().map(|i| (i + 1).encode(env).as_c_arg());
-  unsafe { Term::new(env, list::make_list_from_end(env.as_c_arg(), &mut iter)) }
+    let mut iter = [1, 2, 3].iter().map(|i| (i + 1).encode(env).as_c_arg());
+    unsafe { Term::new(env, list::make_list_from_end(env.as_c_arg(), &mut iter)) }
 }