In particular, allow lists to be built from
double ended iterators, which avoids allocating
intermediate vectors.