@filmor 2022-09-11
Nice! Did you compare the performance of both approaches?