@josevalim 2022-09-11
make_list_from_end was actually 20% faster when encoding arrow arrays. I am yet to benchmark if it makes the default Vector encoding faster too. Later today I Will have some numbers. :)