@josevalim 2022-09-11
Ok, so this is tricky. Here are the numbers:

```
## New (this PR)

Name               ips        average  deviation         median         99th %
16 elems        1.03 M      971.63 ns  ±1498.29%         958 ns        1083 ns
32 elems      668.49 K        1.50 μs   ±918.15%        1.46 μs        1.58 μs
128 elems     220.57 K        4.53 μs   ±388.58%        4.38 μs        4.75 μs
1k elems       31.51 K       31.73 μs     ±5.36%       31.29 μs       36.92 μs
1M elems         28.28       35.37 ms     ±7.33%       33.73 ms       40.14 ms

## Old (master)

Name               ips        average  deviation         median         99th %
16 elems      915.15 K        1.09 μs  ±1399.89%        1.08 μs        1.21 μs
32 elems      696.04 K        1.44 μs   ±943.38%        1.42 μs        1.54 μs
128 elems     274.22 K        3.65 μs   ±289.96%        3.54 μs        3.88 μs
1k elems       42.16 K       23.72 μs    ±67.46%       23.17 μs       29.50 μs
1M elems         38.43       26.02 ms     ±7.68%       24.69 ms       30.16 ms
```

You can see this PR is faster for small lists but then it gets slower, most likely because `enif_make_list_from_array` allocates the whole area upfront, while this PR has to allocate cell by cell. On the other hand, the current approach uses less memory.

I will be glad to revert the changes to `rustler::types::list` if you would prefer to not change the default encoding.