@hansihe 2022-09-11
With regards to https://github.com/erlang/otp/issues/6293:

There are a couple of places I can think of where the performance difference might come from between approach 1. and 2.:

a. The alloc/free of the buffer used to store terms passed to `enif_make_list_from_array`
b. The allocation on the process heap inside `enif_make_list_from_array`/`enif_make_list_cell`
c. The function call overhead each time `enif_make_list_cell` is called

The two different approaches are affected by the different factors as follows:
1. a and b dominates
2. b and c dominates

I have nothing hard to base this on, but I have a gut feeling the reason why approach 2 gets slower is because c becomes the dominating factor.

If this is the case, introducing an API like `enif_alloc_list`/`enif_set_list` might not improve things by much.