@josevalim 2022-09-11
@hansihe if the NIF cost is the culprit, then we can emulate it? Here is what I did, I changed encode to perform a NIF call to `enif_make_list(env, 0)`, which is as cheap as a NIF can be:

```rust
impl<T> Encoder for [T]
where
    T: Encoder,
{
    fn encode<'b>(&self, env: Env<'b>) -> Term<'b> {
        let env_as_c_arg = env.as_c_arg();
        let term_array: Vec<NIF_TERM> = self.iter().map(|x| {
            unsafe { rustler_sys::_enif_make_list(env_as_c_arg, 0) };
            x.encode(env).as_c_arg()
        }).collect();
        unsafe { Term::new(env, list::make_list(env.as_c_arg(), &term_array)) }
    }
}
```

These are the results:

```
16 elems       945.72 K        1.06 μs  ±1417.14%        1.04 μs        1.17 μs
32 elems       703.49 K        1.42 μs   ±889.53%        1.38 μs        1.58 μs
128 elems      278.42 K        3.59 μs   ±367.05%        3.46 μs        3.88 μs
1k elems        43.41 K       23.04 μs    ±51.07%       22.46 μs       29.45 μs
1M elems          37.20       26.88 ms     ±7.57%       25.52 ms       31.26 ms
```

As you can see, it is slightly lower for the large case, but sometimes even slightly faster, so I don't think the NIF call adds to much? WDYT?