@hansihe 2022-09-11
@josevalim Smart way of testing it!

Seems I was very off target with what cost dominated. The API you suggested would probably work pretty well